+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Fix: .CSV files made Excel 2013 compliant 
 * Fix: misbehavour related to 'NOT' and '=' predecession
 * Fix: alasql running from cordova on iOS
+* Fix: SELECT can use functions from GROUP BY list
 
 ## Version 0.2.0 "Athens" (13.07.2015)
 The purpose of this release were hard work on: